@@ -2,23 +2,19 @@
 name = "regtech-user-fi-management"
 version = "0.1.0"
 description = ""
-authors = ["lchen-2101 <73617864+lchen-2101@users.noreply.github.com>"]
+authors = ["CFPB's RegTech Team"]
 readme = "README.md"
-packages = [{ include = "regtech-user-fi-management" }]
+packages = [{ include = "regtech_user_fi_management", from = "src" }]
 
 [tool.poetry.dependencies]
-python = "^3.11"
-fastapi = "^0.109.1"
+python = ">=3.12,<4"
 uvicorn = "^0.22.0"
 python-dotenv = "^1.0.0"
 python-keycloak = "^3.0.0"
 SQLAlchemy = "^2.0.16"
 psycopg2-binary = "^2.9.6"
-python-jose = "^3.3.0"
-requests = "^2.31.0"
-asyncpg = "^0.27.0"
+asyncpg = "^0.29.0"
 alembic = "^1.12.0"
-pydantic-settings = "^2.0.3"
 regtech-api-commons = {git = "https://github.com/cfpb/regtech-api-commons.git"}
 
 [tool.poetry.group.dev.dependencies]
