@@ -4,9 +4,9 @@ from sqlalchemy.ext.asyncio import (
     async_scoped_session,
 )
 from asyncio import current_task
-from config import settings
+from regtech_user_fi_management.config import settings
 
-engine = create_async_engine(settings.inst_conn.unicode_string(), echo=True).execution_options(
+engine = create_async_engine(str(settings.inst_conn), echo=True).execution_options(
     schema_translate_map={None: settings.inst_db_schema}
 )
 SessionLocal = async_scoped_session(async_sessionmaker(engine, expire_on_commit=False), current_task)
